DX9: Fix many compiler warnings on 64bit
- Fixed hundreds of cases of using %x instead of %p format specifiers
  for pointers. 90% of these were done via a sed script.

- Change some types to avoid type cast warnings.

- Add explicit casting to suppress type cast warnings.

- Switch all QueryInterface debug prints to use NameFromIID(riid)
  instead of printing the IID in hex and use %p for the this pointer.

+++ D3D_Shaders/Assembler.cpp
@@ -3340,7 +3340,7 @@ vector<byte> assemblerDX9(vector<char> *asmFile, vector<byte> origBytecode)
 {
 	vector<byte> ret;
 	LPD3DXBUFFER pAssembly;
-	HRESULT hr = D3DXAssembleShader(asmFile->data(), asmFile->size(), NULL, NULL, 0, &pAssembly, NULL);
+	HRESULT hr = D3DXAssembleShader(asmFile->data(), (UINT)asmFile->size(), NULL, NULL, 0, &pAssembly, NULL);
 	if (!FAILED(hr)) {
 		size_t size = pAssembly->GetBufferSize();
 		LPVOID buffer = pAssembly->GetBufferPointer();
